+++ previous-versions/sync-for-commerce-version-1/pkg/models/operations/createconnection.go
@@ -9,7 +9,8 @@ import (
 
 type CreateConnectionRequest struct {
 	RequestBody *string `request:"mediaType=application/json"`
-	CompanyID   string  `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
 }
 
 func (o *CreateConnectionRequest) GetRequestBody() *string {
@@ -28,9 +29,12 @@ func (o *CreateConnectionRequest) GetCompanyID() string {
 
 type CreateConnectionResponse struct {
 	// Success
-	Connection  *shared.Connection
+	Connection *shared.Connection
+	// HTTP response content type for this operation
 	ContentType string
-	StatusCode  int
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
 	RawResponse *http.Response
 }