+++ previous-versions/sync-for-commerce-version-1/pkg/models/shared/billedtotype.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 )
 
+// BilledToType - Defines if the bill or bill credit note is billed/rebilled to a project.
 type BilledToType string
 
 const (