+++ previous-versions/sync-for-commerce-version-1/pkg/models/shared/grouping.go
@@ -2,9 +2,32 @@
 
 package shared
 
+import (
+	"github.com/codatio/client-sdk-go/previous-versions/sync-for-commerce-version-1/pkg/utils"
+)
+
 type Grouping struct {
-	GroupingLevels *GroupingLevels `json:"groupingLevels,omitempty"`
-	GroupingPeriod *GroupingPeriod `json:"groupingPeriod,omitempty"`
+	AdditionalProperties map[string]interface{} `additionalProperties:"true" json:"-"`
+	GroupingLevels       *GroupingLevels        `json:"groupingLevels,omitempty"`
+	GroupingPeriod       *GroupingPeriod        `json:"groupingPeriod,omitempty"`
+}
+
+func (g Grouping) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(g, "", false)
+}
+
+func (g *Grouping) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &g, "", false, false); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (o *Grouping) GetAdditionalProperties() map[string]interface{} {
+	if o == nil {
+		return nil
+	}
+	return o.AdditionalProperties
 }
 
 func (o *Grouping) GetGroupingLevels() *GroupingLevels {