+++ previous-versions/sync-for-commerce-version-1/pkg/models/shared/localization.go
@@ -3,6 +3,7 @@
 package shared
 
 type Localization struct {
+	// `True` if required.
 	Required *bool `json:"required,omitempty"`
 	// Value of the property.
 	Text *string `json:"text,omitempty"`