+++ previous-versions/sync-for-commerce-version-1/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [go v0.20.0] previous-versions/sync-for-commerce-version-1
 ### Releases
 - [Go v0.20.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-commerce-version-1/v0.20.0 - previous-versions/sync-for-commerce-version-1
+
+## 2023-10-09 01:24:09
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce-v1.yaml
+- Speakeasy CLI 1.96.1 (2.150.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.21.0] previous-versions/sync-for-commerce-version-1
+### Releases
+- [Go v0.21.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-commerce-version-1/v0.21.0 - previous-versions/sync-for-commerce-version-1